backend/pyproject.toml
[project]
name = "summa-vision-backend"
version = "0.1.0"
description = "Summa Vision – Canadian housing data ETL and visualization API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.27.0,<1.0.0",
    "pydantic>=2.0,<3.0.0",
    "pydantic-settings>=2.0,<3.0.0",
    "httpx>=0.27.0,<1.0.0",
    "structlog>=24.1.0,<25.0.0",
    "pandas>=2.0.0,<3.0.0",
    "pytz>=2024.1",
    "tzdata>=2024.1",
    "aiobotocore>=2.12.0,<3.0.0",
    "playwright>=1.40.0,<2.0.0",
    "playwright-stealth>=1.0.6,<2.0.0",
    "beautifulsoup4>=4.12.0,<5.0.0",
    "html5lib>=1.1,<2.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13.0,<2.0.0",
    "aiosqlite>=0.20.0,<1.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "greenlet>=3.0.0",
    "apscheduler>=3.10.0,<4.0.0",
    "google-genai>=1.0.0,<2.0.0",
    "cachetools>=5.3.0,<6.0.0",
    "plotly>=5.18.0,<6.0.0",
    "kaleido==0.1.0.post1",
    "Pillow>=10.0.0,<12.0.0",
    "CairoSVG>=2.7.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23,<1.0",
    "respx>=0.21.0,<1.0.0",
    "pytest-cov>=4.0.0,<6.0.0",
    "mypy>=1.8.0,<2.0.0",
    "black>=24.0.0,<25.0.0",
    "types-aiobotocore-s3>=2.12.0",
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_explicit = true

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:kaleido",
    "ignore::DeprecationWarning:pytest_asyncio",
    "ignore::DeprecationWarning:google",
    "ignore::pytest.PytestCollectionWarning",
]
